fix: mistake, forgot

+++ setup-project.sh
@@ -45,7 +45,7 @@ if [ -d "frontend" ]; then
         cd frontend
         cp .env.example .env
         cd ..
-    else [-f "frontend/.example.env"]; then
+    elif [ -f "frontend/.example.env" ]; then
         echo "🎨 setting up frontend env"
         cd frontend
         cp .example.env .env
